fix: address PR review issues for Windows AppContainer sandbox
- Fix exit code truncation: use std::process::exit() with explicit RAII
  drops instead of ExitCode::from(exit_code as u8)
- Extract shared types into sandbox-types crate to eliminate duplication
  between pi and windows-sandbox
- Replace magic ACL numbers with named constants (FILE_GENERIC_READ, etc.)
- Fix unsafe impl Send safety comment with proper justification
- Deduplicate fnv1a_64 by moving to module level in appcontainer.rs
- Remove unused imports from appcontainer.rs and acl.rs
