+++ tauri/sandbox-helper/Cargo.toml
@@ -14,6 +14,7 @@ serde_json = "1.0"
 base64 = "0.22"
 log = "0.4"
 env_logger = "0.11"
+sandbox-types = { path = "../sandbox-types" }
 
 [target.'cfg(windows)'.dependencies]
 windows = { version = "0.58", features = [
