+++ tauri/sandbox-types/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "sandbox-types"
+version = "0.1.0"
+edition = "2021"
+description = "Shared sandbox profile types for browserx"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
